+++ frontend/app/src/composables/api/session/users.ts
@@ -62,7 +62,7 @@ export const useUsersApi = () => {
   const createAccount = async (
     payload: CreateAccountPayload
   ): Promise<PendingTask> => {
-    const { credentials, premiumSetup } = payload;
+    const { credentials, premiumSetup, initialSettings } = payload;
     const { username, password } = credentials;
 
     const response = await api.instance.put<ActionResult<PendingTask>>(
@@ -72,9 +72,7 @@ export const useUsersApi = () => {
         password,
         premiumApiKey: premiumSetup?.apiKey,
         premiumApiSecret: premiumSetup?.apiSecret,
-        initialSettings: {
-          submitUsageAnalytics: premiumSetup?.submitUsageAnalytics
-        },
+        initialSettings,
         syncDatabase: premiumSetup?.syncDatabase,
         asyncQuery: true
       }),